+++ src/app/beer_garden/api/http/handlers/v1/admin.py
@@ -5,9 +5,12 @@ from brewtils.schema_parser import SchemaParser
 
 from beer_garden.api.http.handlers import AuthorizationHandler
 from beer_garden.garden import local_garden
+from beer_garden.metrics import collect_metrics
 
 
 class AdminAPI(AuthorizationHandler):
+
+    @collect_metrics(transaction_type="API", group="AdminAPI")
     async def patch(self):
         """
         ---