+++ src/app/beer_garden/api/http/handlers/v1/command.py
@@ -3,9 +3,12 @@ from brewtils.models import Operation, System
 
 from beer_garden.api.http.handlers import AuthorizationHandler
 from beer_garden.errors import EndpointRemovedException
+from beer_garden.metrics import collect_metrics
 
 
 class CommandAPI(AuthorizationHandler):
+
+    @collect_metrics(transaction_type="API", group="CommandAPI")
     async def get(self, system_id, command_name):
         """
         ---
@@ -44,6 +47,8 @@ class CommandAPI(AuthorizationHandler):
 
 
 class CommandAPIOld(AuthorizationHandler):
+
+    @collect_metrics(transaction_type="API", group="CommandAPIOld")
     async def get(self, command_id):
         """
         ---
@@ -76,6 +81,8 @@ class CommandAPIOld(AuthorizationHandler):
 
 
 class CommandListAPI(AuthorizationHandler):
+
+    @collect_metrics(transaction_type="API", group="CommandListAPI")
     async def get(self):
         """
         ---