+++ src/app/beer_garden/api/http/handlers/v1/queue.py
@@ -3,9 +3,12 @@ from brewtils.models import Operation, Permissions, Queue, System
 
 from beer_garden.api.http.handlers import AuthorizationHandler
 from beer_garden.garden import local_garden
+from beer_garden.metrics import collect_metrics
 
 
 class QueueAPI(AuthorizationHandler):
+
+    @collect_metrics(transaction_type="API", group="QueueAPI")
     async def delete(self, queue_name):
         """
         ---
@@ -37,6 +40,8 @@ class QueueAPI(AuthorizationHandler):
 
 
 class QueueListAPI(AuthorizationHandler):
+
+    @collect_metrics(transaction_type="API", group="QueueListAPI")
     async def get(self):
         """
         ---
@@ -68,6 +73,7 @@ class QueueListAPI(AuthorizationHandler):
         self.set_header("Content-Type", "application/json; charset=UTF-8")
         self.write(response)
 
+    @collect_metrics(transaction_type="API", group="QueueListAPI")
     async def delete(self):
         """
         ---