+++ src/app/beer_garden/api/http/handlers/v1/request.py
@@ -14,10 +14,13 @@ from beer_garden.api.http.base_handler import future_wait
 from beer_garden.api.http.exceptions import BadRequest, RequestForbidden
 from beer_garden.api.http.handlers import AuthorizationHandler
 from beer_garden.errors import UnknownGardenException
+from beer_garden.metrics import collect_metrics
 from beer_garden.requests import remove_bytes_parameter_base64
 
 
 class RequestAPI(AuthorizationHandler):
+
+    @collect_metrics(transaction_type="API", group="RequestAPI")
     async def get(self, request_id):
         """
         ---
@@ -50,6 +53,7 @@ class RequestAPI(AuthorizationHandler):
         self.set_header("Content-Type", "application/json; charset=UTF-8")
         self.write(response)
 
+    @collect_metrics(transaction_type="API", group="RequestAPI")
     async def patch(self, request_id):
         """
         ---
@@ -133,6 +137,8 @@ class RequestAPI(AuthorizationHandler):
 
 
 class RequestOutputAPI(AuthorizationHandler):
+
+    @collect_metrics(transaction_type="API", group="RequestOutputAPI")
     async def get(self, request_id):
         """
         ---
@@ -180,6 +186,7 @@ class RequestOutputAPI(AuthorizationHandler):
 class RequestListAPI(AuthorizationHandler):
     parser = SchemaParser()
 
+    @collect_metrics(transaction_type="API", group="RequestListAPI")
     async def get(self):
         """
         ---
@@ -402,6 +409,7 @@ class RequestListAPI(AuthorizationHandler):
         self.set_header("Content-Type", "application/json; charset=UTF-8")
         self.write(json.dumps(requests))
 
+    @collect_metrics(transaction_type="API", group="RequestListAPI")
     async def post(self):
         """
         ---
@@ -531,6 +539,7 @@ class RequestListAPI(AuthorizationHandler):
         self.set_header("Content-Type", "application/json; charset=UTF-8")
         self.write(response)
 
+    @collect_metrics(transaction_type="API", group="RequestListAPI")
     async def put(self):
         """
         ---
@@ -601,6 +610,7 @@ class RequestListAPI(AuthorizationHandler):
         self.set_header("Content-Type", "application/json; charset=UTF-8")
         self.write(response)
 
+    @collect_metrics(transaction_type="API", group="RequestListAPI")
     async def delete(self):
         """
         ---