+++ src/app/beer_garden/api/http/handlers/vbeta/runner.py
@@ -4,11 +4,13 @@ from brewtils.models import Operation, Permissions
 from brewtils.schema_parser import SchemaParser
 
 from beer_garden.api.http.handlers import AuthorizationHandler
+from beer_garden.metrics import collect_metrics
 
 
 class RunnerAPI(AuthorizationHandler):
     parser = SchemaParser()
 
+    @collect_metrics(transaction_type="API", group="RunnerAPI")
     async def get(self, runner_id):
         """
         ---
@@ -39,6 +41,7 @@ class RunnerAPI(AuthorizationHandler):
         self.set_header("Content-Type", "application/json; charset=UTF-8")
         self.write(response)
 
+    @collect_metrics(transaction_type="API", group="RunnerAPI")
     async def delete(self, runner_id):
         """
         ---
@@ -72,6 +75,7 @@ class RunnerAPI(AuthorizationHandler):
         self.set_header("Content-Type", "application/json; charset=UTF-8")
         self.write(response)
 
+    @collect_metrics(transaction_type="API", group="RunnerAPI")
     async def patch(self, runner_id):
         """
         ---
@@ -145,6 +149,7 @@ class RunnerAPI(AuthorizationHandler):
 class RunnerListAPI(AuthorizationHandler):
     parser = SchemaParser()
 
+    @collect_metrics(transaction_type="API", group="RunnerListAPI")
     async def get(self):
         """
         ---
@@ -169,6 +174,7 @@ class RunnerListAPI(AuthorizationHandler):
         self.set_header("Content-Type", "application/json; charset=UTF-8")
         self.write(response)
 
+    @collect_metrics(transaction_type="API", group="RunnerListAPI")
     async def patch(self):
         """
         ---