+++ src/app/beer_garden/app.py
@@ -35,7 +35,7 @@ from beer_garden.events.processors import (
 )
 from beer_garden.local_plugins.manager import PluginManager
 from beer_garden.log import load_plugin_log_config
-from beer_garden.metrics import PrometheusServer
+from beer_garden.metrics import PrometheusServer, initialize_elastic_client
 from beer_garden.monitor import MonitorFile
 from beer_garden.plugin import StatusMonitor
 from beer_garden.scheduler import MixedScheduler
@@ -57,6 +57,7 @@ class Application(StoppableThread):
     mp_manager = None
     plugin_log_config_observer: MonitorFile = None
     plugin_local_log_config_observer: MonitorFile = None
+    client = None
 
     def __init__(self):
         super(Application, self).__init__(
@@ -67,6 +68,8 @@ class Application(StoppableThread):
 
     def initialize(self):
         """Actually construct all the various component pieces"""
+
+        initialize_elastic_client("core")
 
         # Setup Replication ID for environment
         beer_garden.replication.get_replication_id()